Player controller and flashlight implementation
Created the scene "ChristianTestScene" to test the player movement
Created the script "MouseLook" to handle mouse inputs for moving player camera
Created the script "PlayerMovement" to handle keyboard inputs for moving the player. Also contains the mechanic of turning the flash light on and off.
Created a prefab "Player" with the current version of the player gameObject I was working with.