Added world item interactions to the player
Created Battery prefab which contains a battery object
Created the WorldItemFacer script so that the sprite its attached to always faces the camera.
Created the PlayerWorldInteractions script to handle physics raycasting for interacting with items in the world, such as batteries.
Modified the Player prefab to include PlayerWorldInteractions.
Modified my test scene in the process of developing these changes.