Victim Dialogue System
Added the victim dialogue system.
Added the victim prefab and its script.
Added the ability for the player to interact with a victim, progress through their dialogue, and then make a choice at the end of the dialogue.
The player's movement and looking capabilities are locked while the dialogue is taking place.